Build 4.0.6963.23544
Price updates from Structures
* IPH will now download prices from public structures for the options selected on the Update Prices tab for Single (Region) Select and Price Profiles if the user has the following scopes added: esi-universe.read_structures.v1, esi-markets.structure_markets.v1

Note, CCP only lists public structures that have no blocks on anyone docking. While you may see prices from a structure, not everyone may have access and the prices will not update for that structure through the price update without manually adding by structure ID (see new structure ID pricing feature). Additionally, anyone can set up a buy order anywhere and ESI will return data on those orders even if you don't have access to information about that structure. These buy order prices will not be reflected in update prices due the unavailability of region and solar system identification without access to that information.

* Users can now import prices on the Update Prices tab directly from a structure if it has a market and they can use it by importing with the Structure ID. The structure ID option is enabled on the Update Prices tab when the user has scopes esi-universe.read_structures.v1, esi-markets.structure_markets.v1 enabled. Then if a user wants to import prices from a structure, they follow the following steps through the Add Structure ID form on the Update Prices page:

1. Make a link of the market hub in the EVE Chat window. You can drag the station name from assets or type the name and select auto-link and then search for 'Station' and hit enter to put it in chat. Note, if you do not have access search will not work.
2. Right click the link and select Copy after entering the chat link.
3. Paste the text into the structure ID textbox - it will format the number for you.
4. Click check for Market. If there is no market available or you do not have access, the Structure ID will not be added.

These structure IDs will be kept in IPH until a user removes them from the View Saved Structures window and users can keep querying data from these markets until they no longer have access or the market/structure shuts down.

Account Management Improvements
* Users can now select their character and copy ESI token and scopes data for use in troubleshooting ESI issues.
* Users can now change the name of the 'Dummy Character' by going to Data menu, and selecting 'Change Dummy Character Name'.
* Updated dummy character skills to include basic industry, science, and mining skills.
* Dummy character skills should now load correctly in the skills viewer.

Alpha Account Support
* Added a setting to allow max alpha skills to be added to the dummy character for use with Alpha accounts. Note, this option is only available when using the dummy character.
* Added a setting for indicating that the user is using an Alpha account. This will result in a 2% increase in the facility usage for blueprint costs when checked. Additionally, when checked, IPH will use the selected character's active skills to calculate costs.
* Added a setting to use Active skills over Trained skills. Useful if looking at a character that may have lapsed Omega.

ESI Improvements
* Implemented ESI error limiting, which should eliminate 420 errors users have seen in the past.
* Updated ESI error messaging and processing to provide more information and fewer, duplicate error messages on ESI errors and issues with character authentication and tokens.
* Added an ESI Status monitor for users to check the current status of used scopes within IPH. This form can be viewed by selecting the Data menu, and View ESI Status. Additionally, on startup ESI will now tell users if there may be an issue with the status of ESI queries, which may result in error messages and data unavailable. Users can choose to suppress these error messages in the settings.

Misc New
* Updated Database and Images to January_Release_1.0_2019
* Implemented anonymous usage tracking using Google Analytics to track usage of IPH. Users can disable anonymous tracking in the program settings.
* Added a checkbox to the BP, BP Viewer, and Manufacturing tabs called 'NPC BPOs', which when checked, will allow users to see only blueprints that are sold by NPC corporations.

Bug Fixes
* Invention calculations will now correctly include all skills required for the invention when determining invention chance.
* The blueprint copy used to invent a T2 item will now be included in Invention Materials.
* T2 Abyssal ammo and several T2 standup rigs will now properly load as T2 items instead of T1.
* The Asset Viewer will now break out assets installed on ships into location (slot, cargo) and Corp assets will now be grouped into Corp hangers. Future updates will include additional categories like asset safety locations.
* The shopping list will no longer crash when double-clicking items to load on the BP tab in the items list or build list.
* Rigs added to Copy and Invention structures will now be applied in blueprint calculations correctly.
* The manufacturing tab will now correctly show the invention cost.
* Updated and optimized processing for Market History in an attempt to correct prices cached after long periods of program use resulting in no data returned on the Manufacturing tab.
* Double-clicking on an item in the manufacturing tab will now load the correct decryptor (when used) and properly check the invention time/cost and copy time/cost on the BP tab. Additionally, IPH, profit, and other prices should match correctly between the forms.
* Fixed an issue that wouldn't allow proper sorting of columns on the manufacturing tab or allow adding all columns to the list. Users may have to reset the columns to default to get the change.
* Fixed an issue where deleting all the characters from Manage Accounts would not load the dummy character.
* Improved the functionality of selecting blueprints on the Blueprint tab from the text search option.
* Settings will now save correctly on the BP viewer.

+++ EVEIndustryJobs.vb
@@ -45,6 +45,7 @@ Public Class EVEIndustryJobs
         Dim CacheDate As Date
         Dim LocationID As Long = 0
         Dim TempDate As Date
+        Dim CorpCharIDs As New List(Of Long)
 
         Dim CDType As CacheDateType
 
@@ -62,23 +63,40 @@ Public Class EVEIndustryJobs
                 If IndyJobs.Count > 0 Then
                     Call EVEDB.BeginSQLiteTransaction()
 
-                    ' Clear out all the industry jobs for the user
-                    SQL = "DELETE FROM INDUSTRY_JOBS WHERE InstallerID = " & CharacterTokenData.CharacterID & " AND JobType = " & CStr(JobType)
+                    ' Clear out all the industry jobs for the user if not a corp lookup
+                    If JobType = ScanType.Personal Then
+                        SQL = "DELETE FROM INDUSTRY_JOBS WHERE InstallerID = " & CharacterTokenData.CharacterID & " AND JobType = " & CStr(JobType)
+                    Else
+                        ' Delete all jobs and reload for corp
+                        SQL = "DELETE FROM INDUSTRY_JOBS WHERE JobType = " & CStr(JobType)
+
+                        ' Also, get the list of character IDs stored in the DB with this corporation and only load those jobs
+                        Dim rsIDs As SQLiteDataReader
+                        DBCommand = New SQLiteCommand("SELECT CHARACTER_ID FROM ESI_CHARACTER_DATA WHERE CORPORATION_ID = " & CStr(ID), EVEDB.DBREf)
+                        rsIDs = DBCommand.ExecuteReader
+
+                        While rsIDs.Read
+                            CorpCharIDs.Add(rsIDs.GetInt64(0))
+                        End While
+
+                        rsIDs.Close()
+
+                    End If
 
                     Call EVEDB.ExecuteNonQuerySQL(SQL)
 
-                    ' Insert industry data
-                    For i = 0 To IndyJobs.Count - 1
-                        ' First make sure it's not already in there
-                        With IndyJobs(i)
-                            ' Insert it
-                            If .location_id = 0 Then
-                                LocationID = .station_id
-                            Else
-                                LocationID = .location_id
-                            End If
+                        ' Insert industry data
+                        For i = 0 To IndyJobs.Count - 1
+                            ' First make sure it's not already in there
+                            With IndyJobs(i)
+                                ' Insert it
+                                If .location_id = 0 Then
+                                    LocationID = .station_id
+                                Else
+                                    LocationID = .location_id
+                                End If
 
-                            If .installer_id = CharacterTokenData.CharacterID Then ' update fields
+                            If JobType = ScanType.Personal Or (CorpCharIDs.Contains(.installer_id) And JobType = ScanType.Corporation) Then ' update fields
                                 SQL = "INSERT INTO INDUSTRY_JOBS (jobID, installerID, facilityID, locationID, activityID, "
                                 SQL &= "blueprintID, blueprintTypeID, blueprintLocationID, outputLocationID, "
                                 SQL &= "runs, cost, licensedRuns, probability, productTypeID, status, duration, "
@@ -121,36 +139,40 @@ Public Class EVEIndustryJobs
 
                             End If
                         End With
-                    Next
+                        Next
 
-                    DBCommand = Nothing
+                        DBCommand = Nothing
 
-                    ' Now look up distinct location ids to find any public upwell structures to update
-                    Dim rsStructure As SQLiteDataReader
-                    Dim StructureIDList As New List(Of Long)
+                        ' Now look up distinct location ids to find any public upwell structures to update
+                        Dim rsStructure As SQLiteDataReader
+                        Dim StructureIDList As New List(Of Long)
 
-                    ' Select facilties only for this character, since others may not have the same rights to this token
-                    SQL = "SELECT DISTINCT facilityID FROM INDUSTRY_JOBS WHERE installerID = " & CStr(ID)
-                    DBCommand = New SQLiteCommand(SQL, EVEDB.DBREf)
-                    rsStructure = DBCommand.ExecuteReader
+                        ' Select facilties only for this character, since others may not have the same rights to this token
+                        SQL = "SELECT DISTINCT facilityID FROM INDUSTRY_JOBS WHERE installerID = " & CStr(ID)
+                        DBCommand = New SQLiteCommand(SQL, EVEDB.DBREf)
+                        rsStructure = DBCommand.ExecuteReader
 
-                    While rsStructure.Read
-                        StructureIDList.Add(rsStructure.GetInt64(0))
-                    End While
+                        While rsStructure.Read
+                            StructureIDList.Add(rsStructure.GetInt64(0))
+                        End While
 
-                    rsStructure.Close()
+                        rsStructure.Close()
 
-                    ' Update all the structures we don't have names for
-                    Call UpdateStructureData(StructureIDList, CharacterTokenData)
+                        ' Update all the structures we don't have names for
+                        ' Add the data
+                        Dim SP As New StructureProcessor
+                        For Each StructureID In StructureIDList
+                            Call SP.UpdateStructureData(StructureID, SelectedCharacter.CharacterTokenData, False)
+                        Next
 
-                    DBCommand = Nothing
-                    rsStructure = Nothing
+                        DBCommand = Nothing
+                        rsStructure = Nothing
 
-                    Call EVEDB.CommitSQLiteTransaction()
-                End If
+                        Call EVEDB.CommitSQLiteTransaction()
+                    End If
 
-                ' Update cache date now that it's all set
-                Call CB.UpdateCacheDate(CDType, CacheDate, ID)
+                    ' Update cache date now that it's all set
+                    Call CB.UpdateCacheDate(CDType, CacheDate, ID)
             End If
         End If
